Working all Operational Wallets an Contracts

+++ main/contracts/1_Fundraising.sol
@@ -14,53 +14,53 @@ contract Fundraising {
     IERC20 private token;
     event Transfer_JCO(address from, address to, uint256 amount);
 
-    event Deposit(address indexed sender, uint amount, uint balance);
+    event Deposit(address indexed sender, uint256 amount, uint256 balance);
     event SubmitTransaction(
         address indexed owner,
-        uint indexed txIndex,
+        uint256 indexed txIndex,
         address indexed to,
-        uint value,
+        uint256 value,
         bytes data
     );
-    event ConfirmTransaction(address indexed owner, uint indexed txIndex);
-    event RevokeConfirmation(address indexed owner, uint indexed txIndex);
-    event ExecuteTransaction(address indexed owner, uint indexed txIndex);
+    event ConfirmTransaction(address indexed owner, uint256 indexed txIndex);
+    event RevokeConfirmation(address indexed owner, uint256 indexed txIndex);
+    event ExecuteTransaction(address indexed owner, uint256 indexed txIndex);
 
     address[] public owners;
     mapping(address => bool) public isOwner;
-    uint public numConfirmationsRequired;
+    uint256 public numConfirmationsRequired;
 
     struct Transaction {
         address to;
-        uint value;
+        uint256 value;
         bytes data;
         bool executed;
-        uint numConfirmations;
+        uint256 numConfirmations;
     }
 
     // mapping from tx index => owner => bool
-    mapping(uint => mapping(address => bool)) public isConfirmed;
+    mapping(uint256 => mapping(address => bool)) public isConfirmed;
 
     Transaction[] public transactions;
 
-    modifier txExists(uint _txIndex) {
+    modifier txExists(uint256 _txIndex) {
         require(_txIndex < transactions.length, "tx does not exist");
         _;
     }
 
-    modifier notExecuted(uint _txIndex) {
+    modifier notExecuted(uint256 _txIndex) {
         require(!transactions[_txIndex].executed, "tx already executed");
         _;
     }
 
-    modifier notConfirmed(uint _txIndex, address msg_sender) {
+    modifier notConfirmed(uint256 _txIndex, address msg_sender) {
         require(!isConfirmed[_txIndex][msg_sender], "tx already confirmed");
         _;
     }
 
     constructor(
         address[] memory _owners,
-        uint _numConfirmationsRequired,
+        uint256 _numConfirmationsRequired,
         address _token
     ) {
         require(_owners.length > 0, "owners required");
@@ -70,7 +70,7 @@ contract Fundraising {
             "invalid number of required confirmations"
         );
 
-        for (uint i = 0; i < _owners.length; i++) {
+        for (uint256 i = 0; i < _owners.length; i++) {
             address owner = _owners[i];
 
             require(owner != address(0), "invalid owner");
@@ -90,10 +90,10 @@ contract Fundraising {
 
     function submitTransaction(
         address _to,
-        uint _value,
+        uint256 _value,
         bytes memory _data
     ) public {
-        uint txIndex = transactions.length;
+        uint256 txIndex = transactions.length;
 
         transactions.push(
             Transaction({
@@ -108,7 +108,7 @@ contract Fundraising {
         emit SubmitTransaction(msg.sender, txIndex, _to, _value, _data);
     }
 
-    function confirmTransaction(uint _txIndex, address msg_sender)
+    function confirmTransaction(uint256 _txIndex, address msg_sender)
         public
         txExists(_txIndex)
         notExecuted(_txIndex)
@@ -121,7 +121,7 @@ contract Fundraising {
         emit ConfirmTransaction(msg_sender, _txIndex);
     }
 
-    function executeTransaction(uint _txIndex)
+    function executeTransaction(uint256 _txIndex)
         public
         txExists(_txIndex)
         notExecuted(_txIndex)
@@ -132,7 +132,7 @@ contract Fundraising {
             transaction.numConfirmations >= numConfirmationsRequired,
             "cannot execute tx"
         );
-        uint256 amount = transaction.value * (10 ** 18);
+        uint256 amount = transaction.value * (10**18);
         uint256 token_balance = token.balanceOf(address(this));
         require(amount <= token_balance, "token balance is low");
 
@@ -148,7 +148,7 @@ contract Fundraising {
         emit ExecuteTransaction(msg.sender, _txIndex);
     }
 
-    function revokeConfirmation(uint _txIndex, address msg_sender)
+    function revokeConfirmation(uint256 _txIndex, address msg_sender)
         public
         txExists(_txIndex)
         notExecuted(_txIndex)
@@ -167,19 +167,19 @@ contract Fundraising {
         return owners;
     }
 
-    function getTransactionCount() public view returns (uint) {
+    function getTransactionCount() public view returns (uint256) {
         return transactions.length;
     }
 
-    function getTransaction(uint _txIndex)
+    function getTransaction(uint256 _txIndex)
         public
         view
         returns (
             address to,
-            uint value,
+            uint256 value,
             bytes memory data,
             bool executed,
-            uint numConfirmations
+            uint256 numConfirmations
         )
     {
         Transaction storage transaction = transactions[_txIndex];
